+++ apps/web/src/assets/i18n/zh.json
@@ -783,6 +783,8 @@
         "ACTORS": "演员",
         "DIRECTOR": "导演",
         "SIMILAR": "类似影片",
+        "CREW_JOB_DIRECTOR": "导演",
+        "CREW_JOB_CREATOR": "主创",
         "ACTOR_FILMOGRAPHY": "参演作品",
         "ACTOR_FILTER_ALL": "全部",
         "ACTOR_SCOPE_THIS_PORTAL": "当前门户",
