+++ libs/portal/stalker/feature/src/lib/stalker-series-view/stalker-series-view.component.html
@@ -37,7 +37,7 @@
                         <div class="details__cast-strip">
                             @for (
                                 member of serial.info.tmdb_cast;
-                                track member.name
+                                track member.tmdbPersonId ? 'p' + member.tmdbPersonId : 'i' + $index
                             ) {
                                 <span
                                     class="details__cast-chip"
@@ -83,7 +83,7 @@
                     <span class="label">{{ 'XTREAM.DIRECTOR' | translate }}</span>
                     @if (serial.info.tmdb_directors?.length) {
                         <div class="details__cast-strip">
-                            @for (member of serial.info.tmdb_directors; track member.name) {
+                            @for (member of serial.info.tmdb_directors; track member.tmdbPersonId ? 'p' + member.tmdbPersonId : 'i' + $index) {
                                 <span
                                     class="details__cast-chip"
                                     [class.details__cast-chip--clickable]="
