+++ libs/services/src/lib/tmdb/tmdb-person.ts
@@ -17,19 +17,30 @@ export interface ActorProfile {
     photoUrl: string | null;
 }
 
+export type ActorCrewJob = 'Director' | 'Creator';
+
 export interface ActorFilmographyCredit {
     tmdbId: number;
     mediaType: 'movie' | 'tv';
     title: string;
     year: number | null;
     posterUrl: string | null;
-    /** Character (acting credits) or job like "Director" (crew credits) */
+    /** Character name for acting credits, null for crew-only credits */
     character: string | null;
+    /**
+     * Crew role for directing-only credits — kept separate from
+     * `character` so the UI can render it through translated labels
+     * instead of TMDB's raw English job string.
+     */
+    crewJob: ActorCrewJob | null;
 }
 
 const MAX_FILMOGRAPHY_CREDITS = 80;
 /** Crew jobs worth showing on the person page (directors, TV creators) */
-const FILMOGRAPHY_CREW_JOBS = new Set(['Director', 'Creator']);
+const FILMOGRAPHY_CREW_JOBS = new Set<string>([
+    'Director',
+    'Creator',
+] satisfies ActorCrewJob[]);
 
 export function mapPersonProfile(person: TmdbPersonDetails): ActorProfile {
     return {
@@ -48,7 +59,7 @@ export function mapPersonProfile(person: TmdbPersonDetails): ActorProfile {
  * first (undated entries last), capped at {@link MAX_FILMOGRAPHY_CREDITS}.
  * Acting credits win the dedup so "Actor — Character" survives when a
  * person both starred in and directed the same title; directing-only
- * titles show the job ("Director") in the character slot.
+ * titles carry the crew role in `crewJob` for translated rendering.
  */
 export function mapPersonFilmography(
     person: TmdbPersonDetails
@@ -56,12 +67,14 @@ export function mapPersonFilmography(
     const seen = new Set<string>();
     const credits: ActorFilmographyCredit[] = [];
 
-    const castCredits = person.combined_credits?.cast ?? [];
-    const crewCredits = (person.combined_credits?.crew ?? []).filter(
-        (credit) => credit.job && FILMOGRAPHY_CREW_JOBS.has(credit.job)
+    const castCredits = (person.combined_credits?.cast ?? []).map(
+        (credit) => ({ credit, crewJob: null as ActorCrewJob | null })
     );
+    const crewCredits = (person.combined_credits?.crew ?? [])
+        .filter((credit) => credit.job && FILMOGRAPHY_CREW_JOBS.has(credit.job))
+        .map((credit) => ({ credit, crewJob: credit.job as ActorCrewJob }));
 
-    for (const credit of [...castCredits, ...crewCredits]) {
+    for (const { credit, crewJob } of [...castCredits, ...crewCredits]) {
         const mediaType =
             credit.media_type === 'movie' || credit.media_type === 'tv'
                 ? credit.media_type
@@ -79,7 +92,8 @@ export function mapPersonFilmography(
             title,
             year: extractYear(credit.release_date ?? credit.first_air_date),
             posterUrl: tmdbPosterUrl(credit.poster_path),
-            character: credit.character?.trim() || credit.job?.trim() || null,
+            character: credit.character?.trim() || null,
+            crewJob,
         });
     }
