+++ libs/ui/shared-portals/src/lib/actor-view/actor-view.component.html
@@ -165,6 +165,8 @@
                             }
                             @if (item.character) {
                                 · {{ item.character }}
+                            } @else if (item.crewJob) {
+                                · {{ crewJobKey(item.crewJob) | translate }}
                             }
                         </span>
                     </button>
